# src/main/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Snippets API",
    "description": "Test description",
    "termsOfService": "https://www.google.com/policies/terms/",
    "contact": {
      "email": "contact@snippets.local"
    },
    "license": {
      "name": "BSD License"
    },
    "version": "v1"
  },
  "host": "loco.zerocode.net.cn",
  "schemes": [
    "http"
  ],
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "basic": {
      "type": "basic"
    }
  },
  "security": [
    {
      "basic": []
    }
  ],
  "paths": {
...
    "/apimanager/multiple_delete/": {
      "delete": {
        "operationId": "multiple_delete",
        "summary": "接口数据管理",
        "description": "接口数据管理",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MultiDelete"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "apimanager"
        ]
      },
      "parameters": []
    },
...
    "/asset/multiple_delete/": {
      "delete": {
        "operationId": "multiple_delete",
        "summary": "物料管理",
        "description": "物料管理",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MultiDelete"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "asset"
        ]
      },
      "parameters": []
    },
...
    "/block/multiple_delete/": {
      "delete": {
        "operationId": "multiple_delete",
        "summary": "区块管理",
        "description": "区块管理",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MultiDelete"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "block"
        ]
      },
      "parameters": []
    },
...
    "/datasource/multiple_delete/": {
      "delete": {
        "operationId": "multiple_delete",
        "summary": "在线接口管理",
        "description": "在线接口管理",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MultiDelete"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "datasource"
        ]
      },
      "parameters": []
    },
...
    "/datatype/multiple_delete/": {
      "delete": {
        "operationId": "multiple_delete",
        "summary": "数据类型管理",
        "description": "数据类型管理",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MultiDelete"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "datatype"
        ]
      },
      "parameters": []
    },
...
    "/onlineapi/multiple_delete/": {
      "delete": {
        "operationId": "multiple_delete",
        "summary": "在线接口管理",
        "description": "在线接口管理",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MultiDelete"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "onlineapi"
        ]
      },
      "parameters": []
    },
...
    "/schemadata/multiple_delete/": {
      "delete": {
        "operationId": "multiple_delete",
        "summary": "list:查询",
        "description": "",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MultiDelete"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "schemadata"
        ]
      },
      "parameters": []
    },
...
  },
  "definitions": {
...
    "MultiDelete": {
      "type": "object",
      "properties": {
        "keys": {
          "description": "主键列表",
          "type": "array",
          "items": "string"
        }
      }
    }
  },
...
}